Update project structure and configuration
- Added .nvmrc for Node.js version management.
- Updated .gitignore to exclude .vercel directory.
- Removed deprecated index.html and replaced it with app.html for SvelteKit integration.
- Modified jsconfig.json to extend SvelteKit's TypeScript configuration and updated compiler options.
- Updated package.json to include @sveltejs/kit as a dependency.
- Refactored vite.config.js to use sveltekit plugin and removed unnecessary code.
- Created new app.css and updated main.js to import it correctly.
- Adjusted alertStore and gameStore for improved alert handling.
- Updated layout and page components for SvelteKit compatibility.
