+++ ninho-tech/.gitignore
@@ -31,3 +31,6 @@ node_modules
 .env
 .env.*
 !.env.example
+
+
+.vercel
