+++ ninho-tech/.nvmrc
@@ -1,0 +1,1 @@
+v20.9.0
