--- ninho-tech/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Ninho Tech - Ninho de Letras</title>
-    <link rel="icon" type="image/png" sizes="512x512" href="static/favicon-final.png" />
-  </head>
-  <body>
-    <div id="svelte"></div>
-    <script type="module" src="src/main.js"></script>
-  </body>
-</html>
