+++ ninho-tech/jsconfig.json
@@ -1,7 +1,13 @@
 {
+	"extends": "./.svelte-kit/tsconfig.json",
 	"compilerOptions": {
-		"baseUrl": ".",
-	},
-	"include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.svelte", "src/main.js", "src/main.js", "vite-env.d.ts", "vite-env.d.ts"],
-	"checkJs": true
+		"allowJs": true,
+		"checkJs": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true,
+		"resolveJsonModule": true,
+		"skipLibCheck": true,
+		"sourceMap": true,
+		"strict": false
+	}
 }
