+++ ninho-tech/package.json
@@ -4,13 +4,14 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
     "test": "vitest",
     "test:run": "vitest run"
   },
   "devDependencies": {
+    "@sveltejs/kit": "^2.47.2",
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@vitest/ui": "^2.0.0",
     "jsdom": "^25.0.0",
