+++ ninho-tech/src/app.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html lang="pt-BR">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Ninho Tech - Ninho de Letras</title>
+    <link rel="icon" type="image/png" sizes="512x512" href="%sveltekit.assets%/favicon-final.png" />
+    %sveltekit.head%
+  </head>
+  <body data-sveltekit-preload-data="hover">
+    <div style="display: contents">%sveltekit.body%</div>
+  </body>
+</html>
+
