+++ ninho-tech/src/routes/+page.svelte
@@ -1,0 +1,6 @@
+<script>
+  import Page from "./main.svelte";
+</script>
+
+<Page />
+
